% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Temperatures.r
\name{celsius_to_kelvin}
\alias{celsius_to_kelvin}
\title{Convert between temperature units}
\usage{
celsius_to_kelvin(temp_C)
}
\arguments{
\item{temp_C}{A number, ideally a temperature in Celsius degrees.}
}
\value{
A number with the same temperature \code{temp_C} in Kelvin
}
\description{
Convert between temperature units
}
\examples{
celsius_to_kelvin(0)

}
