% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Temperatures.r
\name{fahrenheit_to_celsius}
\alias{fahrenheit_to_celsius}
\title{Convert between temperature units}
\usage{
fahrenheit_to_celsius(temp_F)
}
\arguments{
\item{temp_F}{A number, ideally a temperature in Farhenheit degrees.}
}
\value{
A number with the same temperature \code{temp_F} in Celsius degrees
}
\description{
Convert between temperature units
}
\examples{
fahrenheit_to_celsius(32)
}
