% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Temperatures.r
\name{fahrenheit_to_kelvin}
\alias{fahrenheit_to_kelvin}
\title{Convert between temperature units}
\usage{
fahrenheit_to_kelvin(temp_F)
}
\arguments{
\item{temp_F}{A number, ideally a temperature in Farhenheit degrees.}
}
\description{
Convert between temperature units
}
\examples{
fahrenheit_to_kelvin(32)


}
