% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthlydiff.R
\name{monthlydiff}
\alias{monthlydiff}
\title{Monthly differences}
\usage{
monthlydiff(df, previous30 = FALSE)
}
\arguments{
\item{df}{name of the data frame}

\item{previous30}{if TRUE it compares current temperatures with previous reference period}
}
\value{
plot of differences in temperatures
}
\description{
Monthly differences
}
\examples{
monthlydiff(simtemperatures, previous30=TRUE)
}
