% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection.R
\name{projection}
\alias{projection}
\title{Temperature projection at given depth}
\usage{
projection(
  df,
  depth = "",
  temp = "",
  start = "2000",
  monthspred = "",
  plot = FALSE
)
}
\arguments{
\item{df}{data frame with depth and temperatures columns}

\item{depth}{depth}

\item{temp}{name of the temperature column}

\item{start}{initial year}

\item{monthspred}{number of months to predict}

\item{plot}{if TRUE plots the projection}
}
\value{
forecast of temperatures
}
\description{
Temperature projection at given depth
}
\examples{
projection(simtemperatures, depth = "-80", temp = "temp.present",
           start = "2000", monthspred = "1