% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simtemperatures.R
\docType{data}
\name{simtemperatures}
\alias{simtemperatures}
\title{temperatures data (simulated)}
\format{
A data frame of 864 rows and 7 columns:
\describe{
\item{depth}{Indicates the depth at which temperature is measured}
\item{month}{Month at which the temperature is recorded}
\item{year}{Year when the temperature is recorded}
\item{temp.present}{Temperature}
\item{temp.hist}{Previous period temperature}
\item{ref_hist}{Period of historical reference}
\item{diff}{Difference between temperatures}

}
}
\usage{
simtemperatures
}
\description{
A data frame of simulated sea temperature observations.
}
\keyword{datasets}
